requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-pptx==0.6.23
pydantic==2.6.4
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.56.0
pdfminer.six==20221105
